Add ControlMode and the start of a "flick and sweep" mode
ControlMode is just an interface for classes that convert sensor
readings into music notes.  The musical note stuff will need to be added
later.

I also added the beginnings of a simple control mode that will play
notes when the user flicks the phone.  The pitches of the notes would
probably be determined by the orientation of the phone.